Update BSP library and sample code for Keil AC6
1. Add *.uvprojx to the original Keil folder of each example.
2. Fix build problems on Keil.
3. Fix build problems on IAR.
4. Fix build problems on GCC.

Author: Hsu-Hung Wu

+++ SampleCode/StdDriver/PS2/main.c
@@ -204,7 +204,7 @@ void PS2D_IRQHandler(void)
 
                     g_cmd[0] = 0;
 
-                    printf("(PS2->STATUS).TXEMPTY is (%0x)\n", ((PS2_GET_STATUS() & PS2_STATUS_TXEMPTY_Msk) >> PS2_STATUS_TXEMPTY_Pos));
+                    printf("(PS2->STATUS).TXEMPTY is (%0x)\n", (uint32_t)(((PS2_GET_STATUS() & PS2_STATUS_TXEMPTY_Msk) >> PS2_STATUS_TXEMPTY_Pos)));
 
                     /* Wait Tx ready */
                     if(PS2_GET_STATUS() & PS2_STATUS_TXEMPTY_Msk)
